Divides the carets from actions.
